requirements-versions.txt
amqp
asgiref
attrs
billiard
boto3
botocore
celery
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
coverage
crispy-bootstrap5
cron-descriptor
diff-match-patch
Django
django-admin-interface
django-appconf
django-axes
django-celery-beat
django-celery-email
django-colorfield
django-cors-headers
django-crispy-forms
django-debug-toolbar
django-environ
django-extensions
django-filter
django-guardian
django-health-check
django-import-export
django-js-asset
django-mptt
django-simple-history
django-storages
django-timezone-field
django-widget-tweaks
django_celery_results
djangorestframework
djangorestframework_simplejwt
dnspython
drf-nested-routers
drf-spectacular
drf-yasg
et_xmlfile
factory_boy
Faker
idna
inflection
iniconfig
jmespath
jsonschema
jsonschema-specifications
kombu
Markdown
numpy
openpyxl
packaging
pandas
pillow
pluggy
prompt_toolkit
psycopg2-binary
Pygments
PyJWT
pytest
pytest-cov
pytest-django
python-crontab
python-dateutil
python-decouple
python-dotenv
python-slugify
pytz
PyYAML
redis
referencing
requests
rpds-py
s3transfer
sentry-sdk
six
sqlparse
tablib
text-unidecode
tqdm
tzdata
tzlocal
uritemplate
urllib3
vine
wcwidth
whitenoise